database: Implement Upsert (Room) for better caching strategies (read commit notes)
Since the API provides the primary key by default, in this app we're using it as the Primary Key for the app.
And, our initial journey for user of the app is =
1. User first downloaded the app and opened the app
2. It will retrieve data from the API and cache it into the local data source, making it as the only SSOT (Single Source of Truth) of data in the app.
   And it will emit the cached data into the UI, not directly from the API.

Scenario above is fine, but if we wanted to introduce like scenario below =
1. User close the app
2. And when user open the app again, i (as a dev) wanted the data in the local data source updated with the newest one on the API, or
3. When user swipe to refresh the data (not implemented yet), i (as a dev) also wanted the data in the local updated with newest one.

Okay, so for that to happen, i implemented the default annotation from Room Database called Upsert (Update or Insert).
Basically, it will update the data by comparing it's primary key from both local or remote. If the data has the same primary key both in local and remote,
It will update the local data with the remote data.
Or, if the data from remote doesn't exist in the local, it will insert the data into the local database.

By implementing this Upsert strategies, those plans of scenario will be successfully implemented and doesn't messed up with the data itself because we're also
embedding is user adding the data to it's favorite or not directly into each of the data in the same table.

References =
https://developer.android.com/reference/kotlin/androidx/room/Upsert